# enhancement(topology): Speed up vector startup with big remap transforms