--- a/lib/enrichment/src/tables.rs
+++ b/lib/enrichment/src/tables.rs
@@ -49,6 +49,15 @@ pub struct TableRegistry {
     tables: Arc<ArcSwap<Option<TableMap>>>,
 }
 
+/// Pessimistic Eq implementation for caching purposes
+impl PartialEq for TableRegistry {
+    fn eq(&self, other: &Self) -> bool {
+        Arc::ptr_eq(&self.tables, &other.tables) && Arc::ptr_eq(&self.loading, &other.loading)
+            || self.tables.load().is_none() && other.tables.load().is_none() && self.loading.lock().unwrap().is_none() && other.loading.lock().unwrap().is_none()
+    }
+}
+impl Eq for TableRegistry {}
+
 impl TableRegistry {
     /// Load the given Enrichment Tables into the registry. This can be new tables
     /// loaded from the config, or tables that need to be reloaded because the