--- a/src/config/graph.rs
+++ b/src/config/graph.rs
@@ -1,10 +1,8 @@
 use indexmap::{set::IndexSet, IndexMap};
 use std::collections::{HashMap, HashSet, VecDeque};
+use crate::config::transform::get_transform_outputs;
 
-use super::{
-    schema, ComponentKey, DataType, OutputId, SinkOuter, SourceOuter, SourceOutput, TransformOuter,
-    TransformOutput,
-};
+use super::{schema, ComponentKey, DataType, OutputId, SinkOuter, SourceOuter, SourceOutput, TransformOuter, TransformOutput};
 
 #[derive(Debug, Clone)]
 pub enum Node {
@@ -76,11 +74,7 @@ impl Graph {
                 id.clone(),
                 Node::Transform {
                     in_ty: transform.inner.input().data_type(),
-                    outputs: transform.inner.outputs(
-                        vector_lib::enrichment::TableRegistry::default(),
-                        &[(id.into(), schema::Definition::any())],
-                        schema.log_namespace(),
-                    ),
+                    outputs: get_transform_outputs(transform, id.clone(), schema.log_namespace())
                 },
             );
         }