@Zettroke 2024-05-20
There is weird test failure for `topology::schema::tests::test_expanded_definition`.
For some reason output doesn't get sorted. And I can't reproduce it locally 
![image](https://github.com/vectordotdev/vector/assets/15730710/ddaaa644-510f-492d-9ab7-3678f33475be)