@Zettroke 2024-05-22
Sure thing. I'll make 2 pull requests then.